Fix piece movements and build v1.1 app
The previous program used in v1.0 is that a piece floats when click it and follows to a cursor. It works on a computer, while does not work well on a tablet. It seems this is because there is no cursor in a tablet.
Considering problems above, I made a program that makes a piece move only when dragging.